requirements.txt
streamlit
python-dotenv
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
pydantic>=2
PyPDF2